+++ biahub/cli/estimate_stitch.py
@@ -8,14 +8,14 @@ from iohub import open_ome_zarr
 
 from biahub.analysis.AnalysisSettings import ProcessingSettings, StitchSettings
 from biahub.cli.parsing import input_position_dirpaths, local, output_filepath
-from biahub.cli.stitch import (
+from biahub.cli.utils import model_to_yaml
+from biahub.stitch import (
     cleanup_shifts,
     compute_total_translation,
     consolidate_zarr_fov_shifts,
     estimate_zarr_fov_shifts,
     get_grid_rows_cols,
 )
-from biahub.cli.utils import model_to_yaml
 
 
 def write_config_file(
@@ -47,7 +47,7 @@ def cleanup_and_write_shifts(
     write_config_file(shifts, output_filepath, channel, fliplr, flipud, rot90)
 
 
-@click.command()
+@click.command("estimate-stitch")
 @input_position_dirpaths()
 @output_filepath()
 @click.option(
